[Update] Migrations - Fixed migration issues which prevented it from running completely

+++ database/migrations/2018_09_09_183918_create_episodes_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
     {
         Schema::create(Episode::TABLE_NAME, function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('next_episode_id')->nullable();
-            $table->unsignedInteger('previous_episode_id')->nullable();
+            $table->unsignedBigInteger('next_episode_id')->nullable();
+            $table->unsignedBigInteger('previous_episode_id')->nullable();
             $table->unsignedBigInteger('season_id');
             $table->unsignedBigInteger('tv_rating_id')->nullable();
             $table->unsignedInteger('number');
